# Optimize the string compare of Slice in hash join