## What type of PR is this：
- [ ] bug
- [ ] feature
- [x] enhancement
- [ ] others

## Which issues of this PR fixes ：
<!--
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
_If PR is about `failing-tests or flakes`, please post the related issues/tests in a comment and do not use `Fixes`_*
-->
Fixes #3905 

## Problem Summary(Required) ：
<!-- (Please describe the changes you have made. In which scenarios will this bug be triggered and what measures have you taken to fix the bug?) -->

Slice operator== has two optimization for string compare
* If the length is equal and larger than 16, use SSE4.1
* If the length is small than 16, convert the address to int16/int32/int64 to comparasion

so i use it to replace the manual implement in HashJoin.

When the cardinality is relatively small, string comparison is the main performance bottleneck, so the optimization effect is very good. When the cardinality is relatively large, cache locality is the main factor, so the performance impact is not large.

```
select count(lo_revenue), count(p_size) from lineorder_str join part_str on lo_partkey=p_partkey;
``` 
* cardinality: 1,400,000
* avg string length: 6.2 byte
* before optimization: 31.82s
* after optimization: 29.48s

```
select count(lo_revenue), count(s_city) from lineorder_str join supplier_str on lo_suppkey=s_suppkey;
```
* cardinality: 200,000
* avg string length: 5.4 byte
* before optimization: 15.98s
* after optimization: 14.39s

```
select count(lo_revenue), count(d_year) from lineorder_str join dates on lo_orderdate=d_datekey;
```
* cardinality: 2556
* avg string length: 8 byte
* before optimization: 4.16s
* after optimization: 3.60s

```
select count(lo_revenue), count(d_year) from lineorder_str join dates on repeat(lo_orderdate,10)=repeat(d_datekey,10);
```
* cardinality: 2556
* avg string length: 80 byte
* before optimization: 13.63s
* after optimization: 10.96s

```
select count(lo_revenue), count(s_city) from lineorder_str join supplier_str on repeat(lo_suppkey,5)=repeat(s_suppkey,5);
```
* cardinality: 200,000
* avg string length: 27 byte
* before optimization: 27.85s
* after optimization: 25.09s

```
select count(lo_revenue), count(s_city) from lineorder_str join supplier_str on repeat(lo_suppkey,10)=repeat(s_suppkey,10);
```
* cardinality: 200,000
* avg string length: 54 byte
* before optimization: 32.72s
* after optimization: 29.54s