@mofeiatwork 2022-03-07
Is there any other optimizing point for string compare? As join and predicate already conducted.